requirements.txt
Django==2.2.2
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-pagedown==1.0.6
markdown2==2.3.8
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
